add log for playoff and promote

+++ common/game.py
@@ -611,8 +611,8 @@ class Battle(object):
 
                 self.turns_sequence[lst_index] = chakra['id']
 
-                dec_z = int(int(round(selected_hero[0]['health'])) / int(round(selected_hero[0]['maxHealth'])))
-                chakra['health'] = chakra['health'] * dec_z
+                dec_z = float(int(round(selected_hero[0]['health'])) / int(round(selected_hero[0]['maxHealth'])))
+                chakra['health'] = int(chakra['health'] * dec_z)
 
                 for spell in self.live_spells:
                     if spell['troop'][0]['id'] == selected_hero[0]['id']:
